Issue 1025 (#1043) - fix printf format strings for Windows Clang build
PR#1025 
* first bunch of files
  sprintf formats and casts
* second batch of printf format fixes
* disable unused debug macros
  and leave in place because I'm not sure who was trying to use them.

+++ IccJSON/IccLibJSON/IccProfileJson.cpp
@@ -85,7 +85,7 @@ static IccJson icJsonGetHeaderFlags(icUInt32Number flags)
   icUInt32Number other = flags & ~(icUInt32Number)(icEmbeddedProfileTrue | icUseWithEmbeddedDataOnly |
                                                    icExtendedRangePCS | icMCSNeedsSubsetTrue);
   if (other) {
-    char buf[16]; snprintf(buf, sizeof(buf), "%08x", other);
+    char buf[16]; snprintf(buf, sizeof(buf), "%08x", (unsigned int) other);
     j["VendorFlags"] = buf;
   }
   return j;
